Linked user Id to shoppingList page

+++ src/main/java/org/elevenfiftyconsulting/repositories/NoteRepository.java
@@ -3,6 +3,6 @@ package org.elevenfiftyconsulting.repositories;
 import org.elevenfiftyconsulting.beans.Note;
 import org.springframework.data.repository.CrudRepository;
 
-public interface NoteRepository extends CrudRepository<Note, Integer> {
+public interface NoteRepository extends CrudRepository<Note, Long> {
 
 }
